pyproject.toml
[project]
name = "codefleet"
version = "0.1.0"
description = "AI Coding Agent Fleet Commander — orchestrate parallel Claude Code agents with Elasticsearch"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "elasticsearch[async]>=8.0",
    "claude-agent-sdk>=0.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "click>=8.0",
    "structlog>=24.0",
]

[project.scripts]
codefleet = "src.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
